+++ src/charts/StackedArea.ts
@@ -7,7 +7,7 @@ class StackedArea extends Chart {
 
     constructor(data: any, userConfig: any = {}) {
         super(
-            new SvgStrategyStreamgraph(), //It uses the same strategy than streamgraph. The only difference is the stack.
+            SvgStrategyStreamgraph, //It uses the same strategy than streamgraph. The only difference is the stack.
             data,
             userConfig,
             defaults