+++ src/svg/strategies/SvgStrategyLinechart.ts
@@ -3,12 +3,13 @@ import Lineset from '../components/Lineset';
 import Pointset from '../components/Pointset';
 import Areaset from '../components/Areaset';
 import Legend from '../components/Legend';
-
+import Annotations from '../components/Annotations';
 import Config from '../../Config';
-import SvgChart from '../base/SvgChart';
+import SvgStrategy from '../base/SvgStrategy';
 import { sortByField } from '../../utils/data/sorting';
 import { convertByXYFormat } from '../../utils/data/transforming';
-class SvgStrategyLinechart extends SvgChart {
+
+class SvgStrategyLinechart extends SvgStrategy {
     /**
      * 
      * XY Axes. Horizontal and vertical references
@@ -28,19 +29,16 @@ class SvgStrategyLinechart extends SvgChart {
     
      */
     private lines: Lineset;
-
-
     private markers: Pointset;
-
     private area: Areaset;
-
     private legend: Legend;
-
+    private annotations : Annotations;
 
     constructor() {
         super();
         this.axes = new XYAxes();
         this.lines = new Lineset(this.axes.x, this.axes.y);
+        this.annotations = new Annotations(this.axes.x, this.axes.y);
     }
 
     public draw(data: [{}]) {
@@ -62,9 +60,10 @@ class SvgStrategyLinechart extends SvgChart {
         super.initialize();
         let markerSize = this.config.get('markerSize'),
             areaOpacity = this.config.get('areaOpacity'),
-            legend = this.config.get('legend');
+            legend = this.config.get('legend'),
+            annotations = this.config.get('annotations');
 
-        this.container.add(this.axes).add(this.lines);
+        this.container.add(this.axes).add(this.lines).add(this.annotations);
 
         if (areaOpacity > 0) {
             this.area = new Areaset(this.axes.x, this.axes.y);