+++ src/svg/strategies/SvgStrategyScatterplot.ts
@@ -3,12 +3,13 @@ import Pointset from '../components/Pointset';
 import Legend from '../components/Legend';
 
 import Config from '../../Config';
-import SvgChart from '../base/SvgChart';
+import SvgStrategy from '../base/SvgStrategy';
 import { sortByField } from '../../utils/data/sorting';
 import { convertByXYFormat } from '../../utils/data/transforming';
 import CanvasPointset from "../components/CanvasPointset";
+import Annotations from "../components/Annotations";
 
-class SvgStrategyScatterplot extends SvgChart {
+class SvgStrategyScatterplot extends SvgStrategy {
     /**
      * 
      * XY Axes. Horizontal and vertical references
@@ -38,14 +39,15 @@ class SvgStrategyScatterplot extends SvgChart {
      * @memberOf SvgStrategyScatterplot
      */
     private canvasMarkers: CanvasPointset;
-
     private legend: Legend;
+    private annotations: Annotations;
 
     constructor() {
         super();
         this.axes = new XYAxes();
         this.markers = new Pointset(this.axes.x, this.axes.y);
         this.canvasMarkers = new CanvasPointset(this.axes.x, this.axes.y);
+        this.annotations = new Annotations(this.axes.x, this.axes.y);
     }
 
     public draw(data: [{}]) {
@@ -67,7 +69,7 @@ class SvgStrategyScatterplot extends SvgChart {
         super.initialize();
         let legend = this.config.get('legend');
 
-        this.container.add(this.axes);
+        this.container.add(this.axes).add(this.annotations);
 
         if (this.config.get('canvas')) {
             this.container.add(this.canvasMarkers);